% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{select_columns}
\alias{select_columns}
\title{Select Columns from a Data.Table}
\usage{
select_columns(data, columns_to_select)
}
\arguments{
\item{data}{A data.table from which columns are to be selected.}

\item{columns_to_select}{A character vector containing the names of the columns to be selected.}
}
\value{
A new data.table containing only the selected columns.
}
\description{
The \code{select_columns} function is designed to extract specific columns from a data.table in R.
}
\examples{
# Example 1: Select specific columns from a data.table
data_example <- data.table(ID = 1:5, Name = c("John", "Alice", "Bob", "Eva", "Mike"), Age = c(25, 30, 22, 28, 35))
selected_columns <- select_columns(data_example, c("ID", "Name"))

# Example 2: Select all columns from a data.table (no change)
selected_all_columns <- select_columns(data_example, colnames(data_example))

}
